+++ app/docker/views/networks/networksController.js
@@ -1,6 +1,6 @@
 import _ from 'lodash-es';
-import DockerNetworkHelper from 'Docker/helpers/networkHelper';
-import { isOfflineEndpoint } from '@/portainer/helpers/endpointHelper';
+import DockerNetworkHelper from '@/docker/helpers/networkHelper';
+import { confirmDelete } from '@@/modals/confirm';
 
 angular.module('portainer.docker').controller('NetworksController', [
   '$q',
@@ -12,18 +12,22 @@ angular.module('portainer.docker').controller('NetworksController', [
   'endpoint',
   'AgentService',
   function ($q, $scope, $state, NetworkService, Notifications, HttpRequestHelper, endpoint, AgentService) {
-    $scope.removeAction = function (selectedItems) {
+    $scope.removeAction = async function (selectedItems) {
+      const confirmed = await confirmDelete('您想删除选定的网络吗？');
+      if (!confirmed) {
+        return null;
+      }
       var actionCount = selectedItems.length;
       angular.forEach(selectedItems, function (network) {
         HttpRequestHelper.setPortainerAgentTargetHeader(network.NodeName);
         NetworkService.remove(network.Id)
           .then(function success() {
-            Notifications.success('已成功删除网络', network.Name);
+            Notifications.success('网络已成功删除', network.Name);
             var index = $scope.networks.indexOf(network);
             $scope.networks.splice(index, 1);
           })
           .catch(function error(err) {
-            Notifications.error('失败', err, '无法删除网络');
+            Notifications.error('Failure', err, '无法删除网络');
           })
           .finally(function final() {
             --actionCount;
@@ -33,8 +37,6 @@ angular.module('portainer.docker').controller('NetworksController', [
           });
       });
     };
-
-    $scope.offlineMode = false;
 
     $scope.getNetworks = getNetworks;
 
@@ -61,12 +63,11 @@ angular.module('portainer.docker').controller('NetworksController', [
       };
 
       if ($scope.applicationState.endpoint.mode.agentProxy && $scope.applicationState.endpoint.mode.provider === 'DOCKER_SWARM_MODE') {
-        req.agents = AgentService.agents();
+        req.agents = AgentService.agents(endpoint.Id);
       }
 
       $q.all(req)
         .then((data) => {
-          $scope.offlineMode = isOfflineEndpoint(endpoint);
           const networks = _.forEach(data.networks, (item) => (item.Subs = []));
           if ($scope.applicationState.endpoint.mode.agentProxy && $scope.applicationState.endpoint.mode.provider === 'DOCKER_SWARM_MODE') {
             $scope.networks = groupSwarmNetworksManagerNodesFirst(data.networks, data.agents);
@@ -81,7 +82,7 @@ angular.module('portainer.docker').controller('NetworksController', [
         })
         .catch((err) => {
           $scope.networks = [];
-          Notifications.error('失败', err, '无法检索网络');
+          Notifications.error('Failure', err, '无法检索网络');
         });
     }
