+++ app/portainer/views/tags/tagsController.js
@@ -1,5 +1,7 @@
 import angular from 'angular';
 import _ from 'lodash-es';
+import { confirmDestructive } from '@@/modals/confirm';
+import { buildConfirmButton } from '@@/modals/utils';
 
 angular.module('portainer.app').controller('TagsController', TagsController);
 
@@ -30,14 +32,24 @@ function TagsController($scope, $state, $async, TagService, Notifications) {
   }
 
   async function removeActionAsync(tags) {
+    const confirmed = await confirmDestructive({
+      title: '是否确定？',
+      message: '确定要删除所选的标签吗？',
+      confirmButton: buildConfirmButton('删除', 'danger'),
+    });
+
+    if (!confirmed) {
+      return;
+    }
+
     for (let tag of tags) {
       try {
         await TagService.deleteTag(tag.Id);
 
-        Notifications.success('Tag successfully removed', tag.Name);
+        Notifications.success('标签删除成功', tag.Name);
         _.remove($scope.tags, tag);
       } catch (err) {
-        Notifications.error('失败', err, 'Unable to remove tag');
+        Notifications.error('失败', err, '无法删除标签');
       }
     }
 
@@ -48,11 +60,11 @@ function TagsController($scope, $state, $async, TagService, Notifications) {
     var tagName = $scope.formValues.Name;
     TagService.createTag(tagName)
       .then(function success() {
-        Notifications.success('Tag successfully created', tagName);
+        Notifications.success('标签创建成功', tagName);
         $state.reload();
       })
       .catch(function error(err) {
-        Notifications.error('失败', err, 'Unable to create tag');
+        Notifications.error('Failure', err, '无法创建标签');
       });
   };
 
@@ -62,7 +74,7 @@ function TagsController($scope, $state, $async, TagService, Notifications) {
         $scope.tags = data;
       })
       .catch(function error(err) {
-        Notifications.error('失败', err, 'Unable to retrieve tags');
+        Notifications.error('Failure', err, '无法检索标签');
         $scope.tags = [];
       });
   }
